Remove intermediatery ProjectUpdater, and CircleCI Model
